Moved ContentTypeMiddleware to response_middleware
